+++ src/show_env/migration_sp_ijg/grid_math/grid_step_estimation.py
@@ -5,7 +5,7 @@ from .grid import Grid
 
 def get_step_from_grid(
     grid: Grid,
-) -> int:
+) -> float:
     first_horizontal_position_index_0 = grid[0]
     if all(
         np.array_equal(
@@ -14,15 +14,15 @@ def get_step_from_grid(
         )
         for first_horizontal_position in grid
     ):
-        return 0
+        return 0.0
     for first_horizontal_position, second_horizontal_position in zip(
         grid[:-1], grid[1:]
     ):
         if first_horizontal_position.xy_tuple != second_horizontal_position.xy_tuple:
-            return int(
+            return float(
                 np.linalg.norm(
                     first_horizontal_position.xy_array
                     - second_horizontal_position.xy_array,
                 )
             )
-    return 0
+    return 0.0
