+++ src/show_env/show_user/convex_hull.py
@@ -39,22 +39,23 @@ def two_dimensionnal_cross_product(
 
 
 def tuple_list_to_array_list(
-    tuple_list: List[Tuple[int, int]]
+    tuple_list: List[Tuple[float, float]]
 ) -> List[npt.NDArray[np.float64]]:
     return [np.array(tuple_element) for tuple_element in tuple_list]
 
 
 def array_list_to_tuple_list(
     array_list: List[npt.NDArray[np.float64]],
-) -> List[Tuple[int, int]]:
+) -> List[Tuple[float, float]]:
     return [
-        (int(array_element[0]), int(array_element[1])) for array_element in array_list
+        (float(array_element[0]), float(array_element[1]))
+        for array_element in array_list
     ]
 
 
 def calculate_convex_hull(
-    positions_tuple: List[Tuple[int, int]]
-) -> List[Tuple[int, int]]:
+    positions_tuple: List[Tuple[float, float]]
+) -> List[Tuple[float, float]]:
     """Graham scan implementation."""
     # IMPROVE: make an object with (index,tuple,array)
     positions_array = tuple_list_to_array_list(positions_tuple)
